GasDispersion.jl

GasDispersion.jl is a set of tools for atmospheric dispersion modeling of gaseous releases, such as might occur during an emergency at a chemical plant or more routinely from a stack. This is intended to be the level of disperson modeling used support consequence analysis or QRA such as is described in Lee's Loss Prevention in the Process Industries or the CCPS Guidelines for Consequence Analysis of Chemical Releases.

Installation

GasDispersion.jl can be installed using Julia's built-in package manager. In a Julia session, enter the package manager mode by hitting ], then run the command

pkg> add https://github.com/aefarrell/GasDispersion.jl

Example usage

Worked examples

This is just a brief overview, detailed worked examples are available as jupyter notebooks

This scenario is adapted from CCPS Guidelines for Consequence Analysis of Chemical Releases, CCPS, pg 47.

Suppose we wish to model the dispersion of gaseous propane from a leak from a storage tank, where the leak is from a 10 mm hole that is 3.5 m above the ground and the propane is at 25°C and 4barg. Assume the discharge coefficient $c_{D} = 0.85$

For ambient conditions we assume the atmosphere is dry air at standard conditions of 1atm and 25°C, with a windspeed of 1.5m/s and class F stability (a "worst case" atmospheric stability)

using GasDispersion

propane = Substance(name = :propane,
                    gas_density = 9.7505, # kg/m^3, NIST Webbook
                    liquid_density = 526.13, # kg/m^3, NIST Webbook
                    reference_temp= 298.15, # K
                    reference_pressure= 501325, # Pa
                    boiling_temp = 231.04, # K, NIST Webbook
                    latent_heat = 16.25/44.0956, # J/kg, NIST Webbook
                    gas_heat_capacity = 1.6849,    # J/kg/K, NIST Webbook
                    liquid_heat_capacity = 2.2460) # J/kg/K, NIST Webbook

scn = scenario_builder(propane, JetSource;
       phase = :gas,
       diameter = 0.01, # m
       dischargecoef = 0.85,
       k = 1.15,         # heat capacity ratio, from Crane's
       temperature = T1, # K
       pressure = P1,    # Pa
       height = 3.5,     # m, height of hole above the ground
       duration = 1)     # s, duration of release

This generates a Scenario defined for a gas jet discharging into dry air at standard conditions. Once we have this defined we can determine the concentration at any point downwind of the release point, assuming the release is a continuous plume, using

# returns a callable
pl = plume(scn, GaussianPlume)

pl(x,y,z) # gives the concentration in kg/m^3 at the point x, y, z

where the coordinate system is such that the release point is at x=0, y=0, z=h

Similarly we could model an instantaneous release, assuming all of the mass was released during 1 second, using a "puff" model

# returns a function
pf = puff(scn, GaussianPuff)

p(x,y,z,t) # gives the concentration in kg/m^3 at the point x, y, z and time t

Building Scenarios

A Scenario is a container for all of the information that a model may need to produce a solution. The intention is for the Scenario to be re-usable, so that the user may run the same scenario with multiple models without much difficulty. Models also have specific parameters, those are handled in the model itself.

A scenario_builder function exists to help create valid Scenarios for various standard release scenarios.

GasDispersion.scenario_builder — Function

scenario_builder(substance::Substance, source::SourceModel, atmosphere::Atmosphere)

Builds a scenario given a substance, source model and an atmosphere. If no atmosphere is given defaults to dry air at ambient conditions and class F stability.

GasDispersion.scenario_builder — Method

scenario_builder(substance::Substance, JetSource, atmosphere::Atmosphere; kwargs...)

Returns returns a scenario for a simple jet from a circular hole. The jet can either be a liquid or a gas (in which case it is assumed to be an ideal gas and the jet is isentropic).

References

CCPS, Guidelines for Consequence Analysis of Chemical Releases, American Institute of Chemical Engineers, New York (1999)

Arguments

phase=:liquid: the phase, either :liquid or :gas

dischargecoef::Number=0.61: the discharge coefficient cd

k::Number=1.4: the heat capacity ratio Cp/Cv

diameter::Number: the diameter of the hole, m

height::Number: the height of the hole, m

pressure::Number: the pressure upstream of the jet, Pa

temperature::Number: the temperature upstream of the jet, K

duration::Number: the duration of the leak, s

Plume Models

Plume models are models of continuous, steady-state, releases and are time independent, this includes, for example, emissions from elevated stacks.

GasDispersion.plume — Function

plume(scenario::Scenario, model::PlumeModel)

Runs the plume dispersion model on the given scenario and returns the solution which is callable to give the concentration c(x, y, z[, t])

The concentration is in kg/m³, if model is unspecified, defaults to a simple gaussian plume model.

All model parameters are assumed to be in SI base units (i.e. distances in m, velocities in m/s, mass in kg, etc.)

Gaussian Plumes

A Gaussian plume is a steady state plume defined by a Gaussian distribution in the y and z directions and an exponential decay in the x direction. The dispersion parameters are correlated to the downwind distance.

GasDispersion.plume — Method

plume(::Scenario, GaussianPlume; kwargs...)

Returns the solution to a Gaussian plume dispersion model for the given scenario.

\[c\left(x,y,z\right) = { \dot{m} \over { 2 \pi \sigma_{y} \sigma_{z} u } } \exp \left[ -\frac{1}{2} \left( y \over \sigma_{y} \right)^2 \right] \\ \times \left\{ \exp \left[ -\frac{1}{2} \left( { z -h } \over \sigma_{z} \right)^2 \right] + \exp \left[ -\frac{1}{2} \left( { z + h } \over \sigma_{z} \right)^2 \right] \right\}\]

where the σs are dispersion parameters correlated with the distance x

References

CCPS, Guidelines for Consequence Analysis of Chemical Releases, American Institute of Chemical Engineers, New York (1999)

Arguments

downwash::Bool=false: when true, includes stack-downwash effects

plumerise::Bool=false: when true, includes plume-rise effects using Briggs' model

Simple Jet Plumes

The simple jet plume is a steady state turbulent jet defined by a Gaussian distribution in the y and z directions. It is similar to the Gaussian plume, however in this case the momentum forming the plume comes entirely from the jet and not the ambient windspeed.

GasDispersion.plume — Method

plume(::Scenario, SimpleJet; kwargs...)

Returns the solution to a simple turbulent jet dispersion model for the given scenario.

\[c\left(x,y,z\right) = k_2 c_0 \left( d \over z \right) \sqrt{ \rho_j \over \rho_a } \exp \left( - \left( k_3 { r \over z } \right)^2 \right)\]

where r is the radial distance from the jet centerline. Assumes a circular jet with diameter equal to the jet diameter. Ground-reflection is included by method of images.

References

Long, V.D., "Estimation of the Extent of Hazard Areas Around a Vent" Chem. Process Hazard, II:6 (1963)

Arguments

release_angle::Number=0: the angle at which the jet is released, in radians

k2::Number=6 parameter of the model, default value is recommended by Long

k3::Number=5 parameter of the model, default value is recommended by Long

Britter-McQuaid Model

The Britter-McQuaid model is an empirical correlation for dense plume dispersion. The model generates an interpolation function for the centerline concentration at the downwind distance x.

GasDispersion.plume — Method

plume(::Scenario, BritterMcQuaidPlume)

Returns the solution to the Britter-McQuaid continuous ground level release model for the given scenario.

References

Britter, R.E. and J. McQuaid, Workbook on the Dispersion of Dense Gases HSE Contract Research Report No. 17/1988, 1988

CCPS, Guidelines for Consequence Analysis of Chemical Releases, American Institute of Chemical Engineers, New York (1999)

Puff Models

Puff models are for "instantaneous" releases or other time-dependent releases.

GasDispersion.puff — Function

puff(scenario::Scenario, model::PuffModel)

Runs the puff dispersion model on the given scenario and returns the solution which is callable to give the concentration c(x, y, z, t)

The concentration is in kg/m³, if model is unspecified, defaults to a simple gaussian puff.

All model parameters are assumed to be in SI base units (i.e. distances in m, velocities in m/s, mass in kg, etc.)

Gaussian Puffs

A Gaussian puff model is defined by gaussian distributions in the x, y, and z directions and travels downwind at the ambient windspeed. The dispersion parameters of the gaussians are correlated with the downwind distance of cloud center.

GasDispersion.puff — Method

puff(::Scenario, GaussianPuff)

Returns the solution to a Gaussian puff dispersion model for the given scenario.

\[c\left(x,y,z,t\right) = \dot{m} \Delta t { { \exp \left( -\frac{1}{2} \left( {x - u t } \over \sigma_x \right)^2 \right) } \over { \sqrt{2\pi} \sigma_x } } { { \exp \left( -\frac{1}{2} \left( {y} \over \sigma_y \right)^2 \right) } \over { \sqrt{2\pi} \sigma_y } }\\ \times { { \exp \left( -\frac{1}{2} \left( {z - h} \over \sigma_z \right)^2 \right) + \exp \left( -\frac{1}{2} \left( {z + h} \over \sigma_z \right)^2 \right) } \over { \sqrt{2\pi} \sigma_z } }\]

References

CCPS, Guidelines for Consequence Analysis of Chemical Releases, American Institute of Chemical Engineers, New York (1999)

Integrated Gaussian Puffs

The integrated Gaussian puff model treats the release as a sum of $n$ equally spaced Gaussian puffs, starting at $t = 0$ to $t = \Delta t$. The default behaviour is to take the limit $n \to \infty$.

GasDispersion.puff — Method

puff(::Scenario, IntPuff; kwargs...)

Returns the solution to an integrated Gaussian dispersion model, where the release is modeled as a sequence of Gaussian puffs, for the given scenario.

\[c\left(x,y,z,t\right) = \sum_{i}^{n-1} { {\dot{m} \Delta t} \over n } { { \exp \left( -\frac{1}{2} \left( {x - u \left( t - i \delta t \right) } \over \sigma_x \right)^2 \right) } \over { \sqrt{2\pi} \sigma_x } } { { \exp \left( -\frac{1}{2} \left( {y} \over \sigma_y \right)^2 \right) } \over { \sqrt{2\pi} \sigma_y } }\\ \times { { \exp \left( -\frac{1}{2} \left( {z - h} \over \sigma_z \right)^2 \right) + \exp \left( -\frac{1}{2} \left( {z + h} \over \sigma_z \right)^2 \right) } \over { \sqrt{2\pi} \sigma_z } }\]

where δt is Δt/n, and the σs are dispersion parameters correlated with the distance x

Arguments

n::Integer: the number of discrete gaussian puffs, defaults to infinity

Britter-McQuaid Model

The Britter-McQuaid model is an empirical correlation for dense cloud dispersion. The model generates an interpolation function for the average cloud concentration and the cloud is rendered as a cylinder.

GasDispersion.puff — Method

puff(scenario::Scenario, BritterMcQuaidPuff)

Returns the solution to the Britter-McQuaid instantaneous ground level release model for the given scenario.

References

Britter, R.E. and J. McQuaid, Workbook on the Dispersion of Dense Gases HSE Contract Research Report No. 17/1988, 1988

CCPS, Guidelines for Consequence Analysis of Chemical Releases, American Institute of Chemical Engineers, New York (1999)