GasDispersion.jl

GasDispersion.jl aims to bring together several models for dispersion modeling of chemical releases with a consistent interface. Currently it is very much under development and significant portions of the code and interface are subject to change.

Installation

GasDispersion.jl can be installed using Julia's built-in package manager. In a Julia session, enter the package manager mode by hitting ], then run the command

pkg> add https://github.com/aefarrell/GasDispersion.jl

Example usage

Worked examples

This is just a brief overview, detailed worked examples are available as jupyter notebooks

This scenario is adapted from CCPS Guidelines for Consequence Analysis of Chemical Releases, CCPS, pg 40.

Suppose a leak of a liquid from a storage tank through a hole. At the hole, the liquid pressure is 120kPa, the liquid has a density of 490 kg/m³, and is in thermal equilibrium with the environment. Assume a circular hole with a discharge coefficient of 0.63 and a diameter of 1cm.

For ambient conditions we assume the atmosphere is at standard conditions of 1atm and 25°C, with a windspeed of 1.5m/s and class F stability (a "worst case" atmospheric stability)

Assumptions:

discharge coefficient, 0.63

diameter of the hole, 0.010m

liquid density, 490kg/m3

liquid temperature, 25C or 298.15K

liquid pressure, 120kPA or 120000Pa

ambient pressure, 1atm or 101325Pa

ambient temperature, 25C or 298.15K

density of air at standard conditions, 1.225kg/m3

ambient windspeed, 1.5m/s

Pasquill-Gifford stability F

using GasDispersion

source=JetSource(phase=:liquid, dischargecoef=0.63, diameter=0.01, duration=1,
                 pressure=120000, temperature=298.15, density=490, height=1)

s = scenario_builder(source, Ambient())

Returns a Scenario defined for a liquid jet discharging into the air at standard conditions. Once we have this defined we can determine the concentration at any point downwind of the release point, assuming the release is a continuous plume, using

# returns a function
c = plume(s, GaussianPlume())

c(x,y,z) # gives the concentration in kg/m^3 at the point x, y, z

where the coordinate system is such that the release point is at x=0, y=0, z=h

Similarly we could model an instantaneous release, assuming all of the mass was released during 1 second, using a "puff" model

# returns a function
c = puff(s, GaussianPuff())

c(x,y,z,t) # gives the concentration in kg/m^3 at the point x, y, z and time t

Building Scenarios

A Scenario is a container for all of the information that a model may need to produce a solution. The intention is for the Scenario to be re-usable, so that the user may run the same scenario with multiple models without much difficulty. Models also have specific parameters, those are handled in the model itself.

By default a Scenario can have any field missing, this is because not all models require all fields. Each model then verifies that none of the necessary fields are missing and throws an error otherwise.

A scenario_builder function exists to help create valid Scenarios for various standard release scenarios.

GasDispersion.scenario_builder — Function

scenario_builder(source::SourceModel, atmosphere::Atmosphere)

Builds a scenario given a source model and an atmosphere. If no atmosphere is given defaults to ambient conditions.

GasDispersion.scenario_builder — Method

scenario_builder(source::JetSource(kwargs), atmosphere::Atmosphere)

Returns returns a scenario with a simple jet source from a circular hole. The jet can either be a liquid or a gas, which case it is assumed to be an ideal gas and the jet is isentropic.

Liquid and gas discharge models are per Guidelines for Consequence Analysis of Chemical Release, CCPS, New York (1999)

Arguments

phase=:liquid: the phase, either :liquid or :gas

dischargecoef::Number=0.61: the discharge coefficient cd

k::Number=1.4: the heat capacity ration cp/cv

diameter::Number: the diameter of the hole

pressure::Number: the pressure upstream of the jet

temperature::Number: the temperature upstream of the jet

density::Number: the density of the fluid upstream of the jet

Plume Models

Plume models are models of continuous, steady-state, releases and are time independent, this includes, for example, emissions from elevated stacks.

GasDispersion.plume — Function

plume(scenario::Scenario, model::PlumeModel)

Runs the plume dispersion model on the given scenario and returns a callable giving the concentration of the form c(x, y, z[, t])

If model is unspecified, defaults to gaussian.

All model parameters are assumed to be in SI base units (i.e. distances in m, velocities in m/s, mass in kg, etc.)

Gaussian Plumes

A gaussian plume is a steady state plume defined by a gaussian distribution in the y and z directions and an exponential decay in the x direction. The dispersion parameters are correlated to the downwind distance.

GasDispersion.plume — Method

plume(::Scenario; GaussianPlume(kwargs...))

Generates a gaussian dispersion model for the given scenario and returns a callable giving the concentration of the form c(x, y, z[, t])

Gaussian plume model is per Guidelines for Consequence Analysis of Chemical Release, CCPS, New York (1999)

\[c\left(x,y,z\right) = { \dot{m} \over { 2 \pi \sigma_{y} \sigma_{z} u } } \exp \left[ -\frac{1}{2} \left( y \over \sigma_{y} \right)^2 \right] \\ \times \left\{ \exp \left[ -\frac{1}{2} \left( { z -h } \over \sigma_{z} \right)^2 \right] + \exp \left[ -\frac{1}{2} \left( { z + h } \over \sigma_{z} \right)^2 \right] \right\}\]

where the σs are dispersion parameters correlated with the distance x

Arguments

downwash::Bool=false: when true, includes stack-downwash effects

plumerise::Bool=false: when true, includes plume-rise effects using Briggs' model

Simple Jet Plumes

The simple jet plume is a steady state turbulent jet defined by a gaussian distribution in the y and z directions. It is similar to the gaussian plume, however in this case the momentum forming the plume comes entirely from the jet and not the ambient windspeed.

GasDispersion.plume — Method

plume(scenario::Scenario, SimpleJet(kwargs...))

Generates a simple turbulent jet dispersion model for the given scenario and returns a callable giving the concentration with the form c(x, y, z[, t])

Turbulent jet model is per Long, V.D., "Estimation of the Extent of Hazard Areas Around a Vent", Chem. Process Hazard, II, 6, 1963 ↩

\[c\left(x,y,z\right) = k_2 c_0 \left( d \over z \right) \sqrt{ \rho_j \over \rho_a } \exp \left( - \left( k_3 { r \over z } \right)^2 \right)\]

where r is the radial distance from the jet centerline. Assumes a circular jet with diameter equal to the jet diameter.Ground-reflection is included by method of images.

Arguments

release_angle::Number=0: the angle at which the jet is released, in radians

k2::Number=6 parameter of the model, default value is recommended by Long

k3::Number=5 parameter of the model, default value is recommended by Long

Britter-McQuaid Model

The Britter-McQuaid model is an empirical correlation for dense plume dispersion. The model generates an interpolation function for the centerline concentration at the downwind distance z.

GasDispersion.plume — Method

plume(scenario::Scenario, BritterMcQuaidPlume())

Generates a Britter-McQuaid dispersion model on the given scenario and returns a callable struct giving the centerline concentration of the form c(x, [y, z, t])

Currently only implements the max concentration at a downwind distance x, the other coordinates are ignored.

Puff Models

Puff models are for "instantaneous" releases or other time-dependent releases, this often includes, for example, releases of vapour clouds.

GasDispersion.puff — Function

puff(scenario::Scenario, model::PuffModel)

Runs the puff dispersion model on the given scenario and returns a callable giving the concentration of the form c(x, y, z, t)

If model is unspecified, defaults to gaussian.

All model parameters are assumed to be in SI base units (i.e. distances in m, velocities in m/s, mass in kg, etc.)

Gaussian Puffs

A gaussian puff model is defined by gaussian distributions in the x, y, and z directions and travels downwind at the ambient windspeed. The dispersion parameters of the gaussians are correlated with the downwind distance.

GasDispersion.puff — Method

puff(scenario::Scenario, GaussianPuff())

Generates a gaussian dispersion model on the given scenario and returns a callable giving the concentration of the form c(x, y, z, t)

Gaussian puff model is per Guidelines for Consequence Analysis of Chemical Release, CCPS, New York (1999)

\[c\left(x,y,z,t\right) = { {\dot{m} \Delta t} \over n } { { \exp \left( -\frac{1}{2} \left( {x - u t } \over \sigma_x \right)^2 \right) } \over { \sqrt{2\pi} \sigma_x } } { { \exp \left( -\frac{1}{2} \left( {y} \over \sigma_y \right)^2 \right) } { \sqrt{2\pi} \sigma_y } }\\ \times { { \exp \left( -\frac{1}{2} \left( {z - h} \over \sigma_z \right)^2 \right) + \exp \left( -\frac{1}{2} \left( {z + h} \over \sigma_z \right)^2 \right) } \over { \sqrt{2\pi} \sigma_z } }\]

Integrated Gaussian Puffs

The integrated gaussian puff model treats the release as a sum of $n$ equally spaced gaussian puffs, starting at $t = 0$ to $t = \Delta t$. The default behaviour is to take the limit $n \to \infty$, hence integrated.

GasDispersion.puff — Method

puff(scenario::Scenario, IntPuff(kwargs...))

Generates an integrated gaussian dispersion model, where the release is modeled as a sequence of gaussian puffs, for the given scenario and returns a callable giving the concentration of the form c(x, y, z, t)

\[c\left(x,y,z,t\right) = \sum_{i}^{n-1} { {\dot{m} \Delta t} \over n } { { \exp \left( -\frac{1}{2} \left( {x - u \left( t - i \delta t \right) } \over \sigma_x \right)^2 \right) } \over { \sqrt{2\pi} \sigma_x } } { { \exp \left( -\frac{1}{2} \left( {y} \over \sigma_y \right)^2 \right) } { \sqrt{2\pi} \sigma_y } }\\ \times { { \exp \left( -\frac{1}{2} \left( {z - h} \over \sigma_z \right)^2 \right) + \exp \left( -\frac{1}{2} \left( {z + h} \over \sigma_z \right)^2 \right) } \over { \sqrt{2\pi} \sigma_z } }\]

where δt is Δt/n, and the σs are dispersion parameters correlated with the distance x

Arguments

npuffs::Integer: the number of discrete gaussian puffs, defaults to infinity

Future

In the future I would like to implement more dense gas dispersion models, as well as a more diverse set of release scenarios.