GasDispersion.jl

GasDispersion.jl aims to bring together several models for dispersion modeling of chemical releases with a consistent interface. Currently it is very much under development and significant portions of the code and interface are subject to change.

Installation

GasDispersion.jl can be installed using Julia's built-in package manager. In a Julia session, enter the package manager mode by hitting ], then run the command

pkg> add https://github.com/aefarrell/GasDispersion.jl

Example usage

Worked examples

This is just a brief overview, detailed worked examples are available as jupyter notebooks

Suppose a chemical release of some substance with a release rate of 1kg/s, at a height of 1m. Using some standard engineering estimates we might end up with a release scenario with the following parameters:

using GasDispersion

s = Scenario(
    1.0,   # mass emission rate, kg/s
    10.0,  # release duration, s
    0.25,  # jet diameter, m
    15.67, # jet velocity, m/s
    1.3,   # jet density, kg/m^3
    101325,# release_pressure, Pa
    450,   # release temperature, K
    1.0,   # release height, m
    1.5,   # windspeed, m/s
    1.225, # ambient density, kg/m^3
    101325,# ambient pressure, Pa
    298.15,# ambient temperature, K
    "F",   # pasquill stability class
)

Once we have this defined we can determine the concentration at any point downwind of the release point, assuming the release is a continuous plume, using

# returns a function
c = plume(s)

c(x,y,z) # gives the concentration in kg/m^3 at the point x, y, z

where the coordinate system is such that the release point is at x=0, y=0, z=h

Similarly we could model an instantaneous release, assuming all of the mass was released during 1 second, using a "puff" model

# returns a function
c = puff(s)

c(x,y,z,t) # gives the concentration in kg/m^3 at the point x, y, z and time t

Building Scenarios

A Scenario is a container for all of the information that a model may need to produce a solution. The intention is for the Scenario to be re-usable, so that the user may run the same scenario with multiple models without much difficulty. Models also have specific parameters, those are handled in the model itself.

By default a Scenario can have any field missing, this is because not all models require all fields. Each model then verifies that none of the necessary fields are missing and throws an error otherwise.

Missing docstring.

Missing docstring for Scenario. Check Documenter's build log for details.

A scenario_builder function exists to help create valid Scenarios for various standard release scenarios.

GasDispersion.scenario_builder — Function

scenario_builder(P, T, conditions; <keyword arguments>)

Builds a scenario given a release pressure, temperature, and set of ambient conditions.

Arguments

P::Number, the release pressure in Pa

T::Number, the release temperature in K

conditions::Scenario=ambient, a scenario or dictionary with the ambient conditions

model=:jet, the model, defaults to a jet

phase=:liquid, the phase of the release, defaults to liquid

stability::String, the Pasquill atmospheric stability

windspeed::Number, the windspeed, defaults to a look-up table for stability

kwargs... each model requires additional keyword arguments

Plume Models

Plume models are models of continuous, steady-state, releases and are time independent, this includes, for example, emissions from elevated stacks.

GasDispersion.plume — Function

plume(scenario::Scenario, model::PlumeModel)

Runs the plume dispersion model on the given scenario and returns a callable giving the concentration of the form c(x, y, z[, t])

If model is unspecified, defaults to gaussian.

All model parameters are assumed to be in SI base units (i.e. distances in m, velocities in m/s, mass in kg, etc.)

plume(::Scenario; GaussianPlume(downwash=false, plumerise=false))

Generates a gaussian dispersion model on the given scenario and returns a callable struct giving the concentration of the form c(x, y, z[, t])

\[c\left(x,y,z\right) = {Q \over 2 \pi u \sigma_{y} \sigma_{z} } \exp \left[ -\frac{1}{2} \left( y \over \sigma_{y} \right)^2 \right] \left\{ \exp \left[ -\frac{1}{2} \left( { z -h } \over \sigma_{z} \right)^2 \right] + \exp \left[ -\frac{1}{2} \left( { z + h } \over \sigma_{z} \right)^2 \right] \right\}\]

downwash controls whether or not stack-downwash is included, by default it is not plumerise controls whether or not plume height is adjusted, by default there is no plume rise

plume(scenario::Scenario, SimpleJet(release_angle=0.0, k2=6, k3=5)

Generates a simple turbulent jet dispersion model on the given scenario and returns a callable struct giving the concentration of the form c(x, y, z[, t])

\[c\left(x,y,z\right) = k_2 c_0 \left( d \over z \right) \sqrt{ \rho_j \over \rho_a } \exp \left( - \left( k_2 { y \over x } \right)^2 \right) \left[ \exp \left( - \left( k_2 { (z-h) \over x }\right)^2 \right) + \exp \left( - \left( k_3 { (z+h) \over x }\right)^2 \right) \right]\]

Assumes a circular jet with diameter equal to the jet diameter. Uses a gaussian concentration profile.

release_angle controls the angle at which the jet is released, in radians k2 and k3 are the parameters of the turbulent jet model.

plume(scenario::Scenario, BritterMcQuaidPlume())

Generates a Britter-McQuaid dispersion model on the given scenario and returns a callable struct giving the centerline concentration of the form c(x, [y, z, t])

Currently only implements the max concentration at a downwind distance x, the other coordinates are ignored.

Puff Models

Puff models are for "instantaneous" releases or other time-dependent releases, this often includes, for example, releases of vapour clouds.

GasDispersion.puff — Function

puff(scenario::Scenario, model::PuffModel)

Runs the puff dispersion model on the given scenario and returns a callable giving the concentration of the form c(x, y, z, t)

If model is unspecified, defaults to gaussian.

All model parameters are assumed to be in SI base units (i.e. distances in m, velocities in m/s, mass in kg, etc.)

puff(scenario::Scenario, GaussianPuff())

Generates a gaussian dispersion model on the given scenario and returns a callable struct giving the concentration of the form c(x, y, z, t)

\[c\left(x,y,z,t\right) = { m \over { (2 \pi)^{3/2} \sigma_x \sigma_y \sigma_z } } \exp \left( -\frac{1}{2} \left( {x - ut} \over \sigma_x \right)^2 \right) \exp \left( -\frac{1}{2} \left( {y} \over \sigma_y \right)^2 \right) \times \left[ \exp \left( -\frac{1}{2} \left( {z - h} \over \sigma_z \right)^2 \right) + \exp \left( -\frac{1}{2} \left( {z + h} \over \sigma_z \right)^2 \right)\right]\]

puff(scenario::Scenario, BritterMcQuaidPuff())

Generates a Britter-McQuaid dispersion model on the given scenario and returns a callable struct giving the concentration of the form c(x, y, z, t)

Future

In the future I would like to implement more dense gas dispersion models, as well as a more diverse set of release scenarios.